Move startup logic to App::onStart
